% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarify.R
\name{squarify}
\alias{squarify}
\title{Treemap in base R}
\usage{
squarify(
  X,
  labels = NULL,
  sub_labels = NULL,
  fontSize = 1,
  col = NULL,
  alpha = 1,
  show_val = TRUE,
  show_pct = FALSE,
  text_col = "#2c3e50",
  sub_text_col = "#34495e",
  borderCol = NULL,
  text_font = 1,
  sub_text_font = 1
)
}
\arguments{
\item{X}{numeric vectors of values}

\item{labels}{a character vector specifying the text to be written for each value in X. Default NULL.}

\item{sub_labels}{a character vector specifying the sub-text to be written for each value in X. Default NULL.}

\item{fontSize}{numeric character expansion factor. Default 1}

\item{col}{colors. If colors is longer than x, the coordinates are recycled to the length of labels.}

\item{alpha}{Default 1}

\item{show_val}{Default TRUE.}

\item{show_pct}{Default FALSE.}

\item{text_col}{Default #2c3e50}

\item{sub_text_col}{Default #34495e}

\item{borderCol}{Default NULL}

\item{text_font}{Font type. Default 1. Can be 1, 2 (bold), 3 (italic), 4 (bold-italic)}

\item{sub_text_font}{Font type. Default 1. Can be 1, 2 (bold), 3 (italic), 4 (bold-italic)}
}
\description{
Treemap in base R
}
\details{
Plot treemaps in base R
This function is a direct transpilation of python \href{https://github.com/laserson/squarify}{squarify} package by \href{https://github.com/laserson}{Uri Laserson}
}
\examples{
gdp <- system.file("extdata", "G7_vs_BRCIS_GDP.tsv", package = "squaRify")
gdp = read.delim(file = gdp)
g7 = subset(gdp, consortium == "G7")
squarify(X = g7$GDP_T, labels = g7$country)
}
